module/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metaphor-machine"
version = "0.1.0"
description = "Algorithmic generation of rich, plot-like style descriptions for AI music systems"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Metaphor Machine Contributors"}
]
keywords = [
    "ai-music",
    "suno",
    "producer-ai",
    "prompt-engineering",
    "generative-music",
    "algorithmic-composition"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0",
]
optimization = [
    "deap>=1.4",
    "scipy>=1.10",
    "numpy>=1.24",
]
embedding = [
    "sentence-transformers>=2.2",
    "scikit-learn>=1.3",
]

[project.scripts]
metaphor = "metaphor_machine.cli.main:cli"
mm = "metaphor_machine.cli.main:cli"

[project.urls]
Homepage = "https://github.com/metaphor-machine/metaphor-machine"
Documentation = "https://metaphor-machine.readthedocs.io"
Repository = "https://github.com/metaphor-machine/metaphor-machine"
Issues = "https://github.com/metaphor-machine/metaphor-machine/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/metaphor_machine"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src/metaphor_machine"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
